+++ frontend/src/styles/theme.js
@@ -1,19 +1,21 @@
 /** Try it out at https://ant.design/theme-editor */
 
 const theme = {
-  colorPrimary: '#e01272',
-  colorInfo: '#1fb393',
-  colorSuccess: '#52c41a',
-  colorWarning: '#faad14',
-  colorError: '#ff4d4f',
-  colorTextBase: '#222222',
-  colorBgBase: '#ffffff',
-  colorBgLayout: '#fafafa',
+  colorPrimary: '#edb048',
+  colorInfo: '#ff5d5f',
+  colorSuccess: '#49aa19', // Dark mode
+  colorWarning: '#d89614', // Dark mode
+  colorError: '#a61d24', // Dark mode
+  colorTextBase: '#ffffffdd',
+  colorBgBase: '#000000',
+  colorBgLayout: '#222222',
+  colorBgLayoutDarker: '#181818',
   baseBorderRadius: '6px',
   boxShadow:
     '0 1px 2px 0 rgba(0, 0, 0, 0.03), 0 1px 6px -1px rgba(0, 0, 0, 0.02), 0 2px 4px 0 rgba(0, 0, 0, 0.02)',
   boxShadowSecondary:
     '0 6px 16px 0 rgba(0, 0, 0, 0.08),0 3px 6px -4px rgba(0, 0, 0, 0.12),0 9px 28px 8px rgba(0, 0, 0, 0.05)',
+  fontFamily: 'Cascadia',
 };
 
 export default theme;